+++ energyquantified/api/timeseries.py
@@ -3,9 +3,9 @@ from .base import BaseAPI
 from ..metadata import CurveType
 from ..parser.timeseries import parse_timeseries
 
-
 # Tuple of supported values for Curve.curve_type in the time series API
 CURVE_TYPES = (CurveType.TIMESERIES, CurveType.SCENARIO_TIMESERIES)
+
 
 class TimeseriesAPI(BaseAPI):
     """
@@ -26,7 +26,8 @@ class TimeseriesAPI(BaseAPI):
             end=None,
             frequency=None,
             aggregation=None,
-            hour_filter=None):
+            hour_filter=None,
+            threshold=None):
         """
         Load time series data for a :py:class:`energyquantified.metadata.Curve`.
 
@@ -47,6 +48,10 @@ class TimeseriesAPI(BaseAPI):
         :param hour_filter: Filters on hours to include (i.e. BASE, PEAK),\
             has no effect unless *frequency* is provided, defaults to BASE
         :type hour_filter: Filter, optional
+        :param threshold: Allow that many values to be missing within one frame of \
+            *frequency*. Has no effect unless *frequency* is provided, \
+            defaults to 0.
+        :type threshold: int, optional
         :return: A time series
         :rtype: :py:class:`energyquantified.data.Timeseries`
         """
@@ -61,6 +66,7 @@ class TimeseriesAPI(BaseAPI):
         if "frequency" in params:
             self._add_aggregation(params, "aggregation", aggregation)
             self._add_filter(params, "hour-filter", hour_filter)
+            self._add_int(params, "threshold", threshold, min=0)
         # HTTP request
         response = self._get(url, params=params)
         return parse_timeseries(response.json())